refactor: refactor BaseTable : make it reuseable. refactor all table: use shared BaseTable.
