+++ src/components/views/DatabaseView/shared/index.ts
@@ -1,3 +1,4 @@
 // 共用組件的統一導出
 export { default as EditableCell } from "./EditableCell";
 export { Filter } from "./Filter";
+export { default as BaseTable } from "./BaseTable";
